+++ CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- links to the app info in web app
+- Links to the app info in web app
+- Loading indicator for search page in web app
 
 ### Fixed